backend/service_auth/requirements.txt
Flask==3.0.3
redis==5.1.1
pyjwt==2.9.0
bcrypt==4.2.0
SQLAlchemy==2.0.35
mysql-connector-python==9.1.0